Initial Commit
This is the initial commit. This code has been scaped together from a variety of code segments and sources including partial forum conversations and deprecated Wiki's (No I did not write or invent this code.) The code segments were targeted to work with now deprecated versions of Python, and has since been ported to Python 3 with this release.